phpunit: remove cache AJAX tests to proper AJAX testing class
